[DDW-652]: Split epochs to different components
