+++ source/renderer/app/themes/daedalus/cardano.js
@@ -552,7 +552,7 @@ const CARDANO_THEME_CONFIG = {
     '--theme-staking-font-color-regular': '#5e6066',
     '--theme-staking-font-color-light': 'rgba(94, 96, 102, 0.7)',
     '--theme-staking-font-color-lighter': 'rgba(94, 96, 102, 0.5)',
-    '--theme-staking-table-head-background-color': '#f1f3f5',
+    '--theme-staking-table-head-background-color': 'rgba(32, 34, 37, 0.07)',
     '--theme-staking-table-border-color': '#c6cdd6',
     '--theme-staking-link-color': 'rgba(28, 172, 99, 1)',
     '--theme-staking-link-color-light': 'rgba(28, 172, 99, 0.5)',
